SB [feature/validate-words-points] temporary version to put validate words and put letters back on rack

+++ client/src/app/services/validation.service.ts
@@ -29,70 +29,35 @@ export class ValidationService {
     /* eslint-enable no-unused-vars */
 
     updatePlayerScore(newWords: ScrabbleWord[], player: Player): void {
-        const wordsValue = this.validateWordsAndCalculateScore(newWords);
+        const wordsValue = this.calculateScore(newWords);
         player.score += wordsValue;
         // Retirer lettres du board
         setTimeout(() => {
-            newWords.forEach((newWord) => {
-                // for (let j = 0; j < newWord.content.length; j++) {
-                //     if (wordsValue === 0) {
-                //         if (newWord.orientation === WordOrientation.Vertical) {
-                //             this.gridService.removeSquare(newWord.startPosition.x, newWord.startPosition.y + j);
-                //         }
-                //         if (newWord.orientation === WordOrientation.Horizontal) {
-                //             this.gridService.removeSquare(newWord.startPosition.x + j, newWord.startPosition.y);
-                //         }
-                //     } else {
-                //         // TODO: are all the right squares set to isvalidated? 
-                //         if (newWord.orientation === WordOrientation.Vertical) {
-                //             this.gridService.scrabbleBoard.squares[newWord.startPosition.x][newWord.startPosition.y + j].isValidated = true;
-                //         }
-                //         if (newWord.orientation === WordOrientation.Horizontal) {
-                //             this.gridService.scrabbleBoard.squares[newWord.startPosition.x + j][newWord.startPosition.y].isValidated = true;
-                //         }
-                //         this.bonusService.useBonus(newWord);
-                //     }
-                // }
-                for(let letter of newWord.content){
-                    // TODO: technically if we only put white letters the value would be 0 but it should be valid.
-                    // if we separate validateWords and CalculateScore, we could juste use validateWords' return (true/false)
-                    // the place method calls validateWordsAndCalculateScore and then calls updatePlayerScore
-                    // instead it could just call validate words and then this methods calls calculateScore and yatti yatta
-                    if(wordsValue === 0){
-                        this.gridService.removeSquare(letter.tile.position.x,letter.tile.position.y);
-                    }else{
-                        // TODO: seems to have a problem in how usebonus iterates? did the same in previous for loop's version that is commented
-                        // where it also put bonus used and isValided true to empty squares after the placed word
-                        this.bonusService.useBonus(newWord); 
+            if (this.validateWords(newWords)) {
+                newWords.forEach((newWord) => {
+                    // actually what was this method supposed to do? seemmed to do a few things at the same time
+                    for (const letter of newWord.content) {
+                        if (wordsValue === 0) {
+                            this.gridService.removeSquare(letter.tile.position.x, letter.tile.position.y);
+                        } else {
+                            // if change the isvalidated = true here, change how its used in solo game service
+                            this.bonusService.useBonus(newWord);
+                        }
                     }
-                }
-
-            });
+                });
+            }
         }, WAIT_TIME);
     }
 
-    // TODO: separate validate words and calculate score in 2 methods?
-    validateWordsAndCalculateScore(newWords: ScrabbleWord[]): number {
+    calculateScore(newWords: ScrabbleWord[]): number {
         let totalScore = 0;
 
-        for (let i = 0; i < newWords.length; i++) {
-            this.words[i] = this.convertScrabbleWordToString(newWords[i].content);
-            // Word not valid, validation fails3
-            if (!this.isWordValid(this.words[i])) {
-                return 0;
-            } else {
-                // Words are all valid
-                // Add word's value to player's score
-                newWords[i].value = this.bonusService.totalValue(newWords[i]);
-                totalScore += newWords[i].value;
-
-                // TODO: is it fine to put this here? if put in updatePlayerScore, because of how the methods are called in placer, it wouldnt work properly
-                newWords[i].content.forEach((letter) => {
-                    letter.tile.isValidated = true;
-                });
-
-            }
+        // Adding total score for each new word
+        for (const word of newWords) {
+            word.value = this.bonusService.totalValue(word);
+            totalScore += word.value;
         }
+
         if (this.newLettersCount() === BONUS_LETTER_COUNT) {
             // Add 50 points to player's score
             totalScore += BONUS_POINTS;
@@ -100,6 +65,22 @@ export class ValidationService {
             return 0;
         }
         return totalScore;
+    }
+
+    validateWords(newWords: ScrabbleWord[]): boolean {
+        for (let i = 0; i < newWords.length; i++) {
+            this.words[i] = this.convertScrabbleWordToString(newWords[i].content);
+            // Word not valid, validation fails3
+            if (!this.isWordValid(this.words[i])) {
+                return false;
+            } else {
+                // Word was valid, set its letters as validated
+                newWords[i].content.forEach((letter) => {
+                    letter.tile.isValidated = true;
+                });
+            }
+        }
+        return true;
     }
 
     // Total value ne consume pas les bonus
